+++ speeduino/sensors.h
@@ -29,7 +29,7 @@ volatile int AnChannel[15];
 unsigned long MAPrunningValue; //Used for tracking either the total of all MAP readings in this cycle (Event average) or the lowest value detected in this cycle (event minimum)
 unsigned long EMAPrunningValue; //As above but for EMAP
 unsigned int MAPcount; //Number of samples taken in the current MAP cycle
-uint16_t MAPcurRev; //Tracks which revolution we're sampling on
+uint32_t MAPcurRev; //Tracks which revolution we're sampling on
 
 //These variables are used for tracking the number of running sensors values that appear to be errors. Once a threshold is reached, the sensor reading will go to default value and assume the sensor is faulty
 byte mapErrorCount = 0;
